+++ src/phonetics/french/fonem.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 /**
  * Talisman phonetics/french/fonem
  * ================================
@@ -34,7 +33,7 @@ const RULES = {
   // Vowels & vowel clusters
   'V-1': [/E?AU/g, 'O'],
   'V-2': [/(?:E?AU|O)LT$/, 'O'],
-  'V-3': [/E?AU$/, 'O'],
+  'V-3': [/E?AUT$/, 'O'],
   'V-4': [/E?AUX$/, 'O'],
   'V-5': [/(?:E?AU|O)LX$/, 'O'],
   'V-6': [/E?AUL?D$/, 'O'],
@@ -42,36 +41,37 @@ const RULES = {
   'V-8': [/EUX$/, 'EU'],
   'V-9': [`EY(?=$|${C})`, 'E'],
   'V-10': [`(${C})?Y(?!${V})`, '$1I'],
-  'V-11': [`(${V})ILL`, '$1Y'],
-  'V-12': [`OU(?=I(?!LL)|[AEOU])`, 'W'],
-  'V-13': [`(${V})\\1`, '$1'],
+  'V-11': [`(${V})I(?=${V})`, '$1Y'],
+  'V-12': [`(${V})ILL`, '$1Y'],
+  'V-13': [/OU(?=I(?!LL)|[AEOU])/g, 'W'],
+  'V-14': [`(${V})\\1`, '$1'],
 
   // Nasals
-  'V-14': [`[AE]M(?=[${CONSONANTS}N](?!$))`, 'EN'],
-  'V-15': [`OM(?=[${CONSONANTS}N])`, 'ON'],
-  'V-16': [`AN(?=${C}))`, 'EN'],
-  'V-17': [`(?:AIM|AIN|EIN)(?=${C})`, 'IN'],
+  'V-15': [`[AE]M(?=[${CONSONANTS}N](?!$))`, 'EN'],
+  'V-16': [`OM(?=[${CONSONANTS}N])`, 'ON'],
+  'V-17': [`AN(?=${C})`, 'EN'],
+  'V-18': [`(?:AIM|AIN|EIN)(?=${C}|$)`, 'IN'],
   'V-19': [/(?:BORNE?|BOURNE?|BURNE)$/, 'BURN'],
   'V-20': [`(?:^IM|IM(?=${C}))(?=[${CONSONANTS}N])`, 'IN'],
 
   // Consonants & consonant clusters
-  'C-1': [/VB/g, 'V'],
+  'C-1': [/BV/g, 'V'],
   'C-2': [`(${V})C(?=[EIY])`, '$1SS'],
-  'C-3': [`([^CX]C(?=[EIY]))`, '$1S'],
+  'C-3': [/([^CX])C(?=[EIY])/g, '$1S'],
   'C-4': [/^C(?=[EIY])/, 'S'],
-  'C-5': [/^C(=[AOU])/, 'K'],
+  'C-5': [/^C(?=[AOU])/, 'K'],
   'C-6': [`(${V})C$`, '$1K'],
   'C-7': [`C(?=[${CONSONANTS}CH])`, 'K'],
   'C-8': [/CC(?=[AOU])/g, 'K'],
   'C-9': [/CC(?=[EIY])/g, 'X'],
-  'C-10': [/G(?=[EIY])/g, 'G'],
-  'C-11': [/GA(?=I?[MN])/g, 'GA'],
+  'C-10': [/G(?=[EIY])/g, 'J'],
+  'C-11': [/GA(?=I?[MN])/g, 'G§'], // The paper is inconsistent so I cheated.
   'C-12': [/(?:GEO|GEAU)/g, 'JO'],
   'C-13': [`GNI(?=${V})`, 'GN'],
-  'C-14': [/H(?![CPS])/g, ''],
+  'C-14': [/(^|[^CPS])H/g, '$1'],
   'C-15': [/JEA/g, 'JA'],
-  'C-16': [`^MAC(?=${C})`, 'MAC'],
-  'C-17': [/^MC/, 'MAC'],
+  'C-16': [`^MAC(?=${C})`, 'M§'], // The paper is inconsistent so I cheated.
+  'C-17': [/^MC/, 'M§'],
   'C-18': [/PH/g, 'F'],
   'C-19': [/QU/g, 'K'],
   'C-20': [/^SC(?=[EIY])/, 'S'],
@@ -79,13 +79,15 @@ const RULES = {
   'C-22': [/(.)SC(?=[AOU])/g, '$1SK'],
   'C-23': [/SH/g, 'CH'],
   'C-24': [/TIA$/, 'SSIA'],
-  'C-25': [`([AIOUY])W`, '$1'],
-  'C-26': [/C(?=[CSZ])/g, 'X'],
-  'C-27': [`(?:Z(?=${V})|(${C})Z${C})`, '$1S'],
-  // Some rules cannot be reduced to regular expressions...
-  'C-30': [/^(?:SIN|SAIN|SEIN|SINT|SEIM|CIN|CINQ|SAINT)/, 'ST-'],
-  'C-31': [/^SAINTE/, 'STE-'],
-  'C-32': [/^ST/, 'ST-'],
+  'C-25': [/([AIOUY])W/g, '$1'],
+  'C-26': [/X[CSZ]/g, 'X'],
+  'C-27': [`(?:Z(?=${V})|(${C})Z(?=${C}))`, '$1S'],
+  'C-28': [`(?:([${CONSONANTS}CLS])\\1|(C)C(?!${V})|(.S)S(?!${V})|([^I]L)L)`, '$1$2$3$4'],
+  'C-28-bis': [/ILE$/, 'ILLE'],
+  'C-29': [`(?:(ILS)|([CS]H)|([MN]P)|(R[CFKLNSX])|(${C})${C})$`, '$1$2$3$4$5'],
+  'C-30': [/^(?:SINT?|SAINT?|SEIN|SEIM|CINQ?)/, 'ST-'],
+  'C-31': [/^SAINTE/, 'STE-'],
+  'C-32': [/^ST(?!E)/, 'ST-'],
   'C-33': [/^STE/, 'STE-']
 };
 
@@ -97,21 +99,32 @@ for (const k in RULES) {
     rule[0] = new RegExp(rule[0], 'g');
 }
 
+// Rules must be ordered
 const ORDERED_RULES = [
-  'V-14', 'C-28',
+  'V-14', 'C-28', 'C-28-bis', // START
   'C-12',
+  'C-9',
   'C-10',
+  'C-16',
+  'C-17',
+  'C-20',
+  'C-2', 'C-3', 'C-7',
   'V-2', 'V-3', 'V-4', 'V-5', 'V-6',
   'V-1',
   'C-14',
   'C-11',
-  'C-30', 'C-31',
+  'C-33', 'C-32', 'C-31', 'C-30', // SAINT
   'V-15', 'V-17', 'V-18',
-  'C-16',
-  'C-17',
-  'C-2', 'C-3', 'C-7',
+  'V-7', 'V-8', 'V-9', 'V-10', 'V-11', 'V-12', 'V-13', 'V-16', 'V-19', 'V-20', // V
+  'C-1', 'C-4', 'C-5', 'C-6', 'C-8', 'C-13', 'C-15', 'C-18', // C
+  'C-19', 'C-21', 'C-22', 'C-23', 'C-24', 'C-25', 'C-26', 'C-27', // C
   'C-29', // END
-  'V-14', 'C-28' // ONCE AGAIN
+  'V-14', 'C-28', 'C-28-bis' // ONCE AGAIN
+].map(key => RULES[key]);
+
+const FIXING_RULES = [
+  [/G§/g, 'GA'],
+  [/M§/g, 'MAC']
 ];
 
 /**
@@ -122,11 +135,24 @@ const ORDERED_RULES = [
  *
  * @throws {Error} The function expects the name to be a string.
  */
-export default function phonetic(name) {
+export default function fonem(name) {
   if (typeof name !== 'string')
     throw Error('talisman/phonetics/french/fonem: the given name is not a string.');
 
   let code = deburr(name)
     .toUpperCase()
     .replace(/[^A-Z]/, '');
+
+  // Applying rules in order
+  for (let i = 0, l = ORDERED_RULES.length; i < l; i++) {
+    const [pattern, replacement] = ORDERED_RULES[i];
+
+    code = code.replace(pattern, replacement);
+  }
+
+  // Fixing rules
+  for (let i = 0, l = FIXING_RULES.length; i < l; i++)
+    code = code.replace(...FIXING_RULES[i]);
+
+  return code;
 }