+++ test/endpoint.js
@@ -93,6 +93,7 @@ describe('phonetics', function() {
   require('./phonetics/statcan.js');
 
   describe('french', function() {
+    require('./phonetics/french/fonem.js');
     require('./phonetics/french/phonetic.js');
     require('./phonetics/french/phonex.js');
     require('./phonetics/french/sonnex.js');